Completed basic scoping.
Instead of a Symbol type as was suggested in the previous commit,
a Scoped interface was created which the Scope object holds on to.
Right now, only Block and Function are scoped, but as more statements
are introduced this will change. At this point, parsing becomes
a valid target and is what will be added now.
Types have also been given makeshift coercibleTo() and convertibleTo()
checker functions. These will be removed soon and replaced with
well-defined casting and coercing rules.