Preparing for scoping rewrite
The Scope and Scoped types need a rewrite to integrate well with the
rest of the system (particularly Variable and Function). As such, at the
same time parsing will be slowly introduced with a Parsable interface,
such that these types can define how they should be parsed individually.
At the same time, a similar introduction for code generation will be
made with a Generatable interface or similar.

These new changes will be made on a new branch so that they can be
compared and differences cleaned out. Additionally, that new development
branch is likely to become the primary branch from which development
occurs rather than the master branch, which is expected to be more
stable.

This commit itself simply cleans up some virtual functions and classes
that were being defined badly. Additionally, the README and LICENSE
files have been moved to where they belong at the project root.